Add comments functionality to posts
This commit introduces the ability for users to leave comments on posts. It includes the creation of a new 'comments' table in the database and corresponding Comment model, as well as related controller and factory. Commenting functionality is made available in the post view, and comments are displayed with the post. User authorization for commenting is also implemented.