Add admin middleware and update post creation process
The admin middleware is introduced to restrict access to certain routes to admins only. In addition, the process of creating posts has been revamped by adding an improved form in the 'create post' page, and implementing storage and validation rules in the PostController. Lastly, a thumbnail field has been added to the 'posts' table in the database.